DataStream/requirements.txt
Django>=5.1.4
mysqlclient>=2.2.1
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
django-environ>=0.11.2
pillow>=10.2.0
django-filter>=23.5
reportlab>=4.1.0
openpyxl>=3.1.2 